+++ test/conftest.py
@@ -1,1 +1,4 @@
-import pytest  # noqa
+from pathlib import Path
+
+
+SPEC_DIR = Path("torcharc/example/spec")
